+++ src/i18n/resources/en/permissions.ts
@@ -28,6 +28,8 @@ export const permissions = {
     notification: "Notifications",
     users: "User management",
     tenant: "Tenants",
+    app: "Applications",
+    "registration-defaults": "Registration defaults",
     system: "System",
   },
 };